Fix hourly stats
Just after midnight, when we compute the daily and hourly stats for the
day that has just passed, we don't have yet the full data for the first hour
of the new day; therefore, we may only compute the hourly stats for
target day + 0 hour, target day + 1 hour, ..., target day + 23 hour, but
*not* for target day + 24 hour.

Otherwise, as we have only 1 or two telegrams, we would already create an entry
for target day + 24 hours (= target day + 1 day + 0 hour) with an incorrect
consumption of 0 (or nearly 0).

On the following day, we would again look at the hour starting at the next
target day, but as the entry already existed, we would not recompute it.

This is the easiest fix for the problem. Alternatively, one could also
- wait for 1 am before computing stats
- change the code to update an already existing hourly cosumption