Further correction for every hourly consumption
As the total hourly consumption is computed by the difference of
the meter readings between the end of the hour and the beginning
(and not by *summing* individual consumptions by minute), we
must consider the reading at [start + 1 hour] itself, therefore
we need to use 'lte' for the upper bound in the filter of
get_consumption_by_range.

The current code is systematically ignoring the consumption of
the last minute of the hour.